requirements.txt
celery==5.3.1
django==4.2.2 ; python_version >= '3.8'
django-cors-headers==4.1.0
djangorestframework~=3.14.0
drf-spectacular~=0.26.2
gunicorn~=20.1.0
pandas==2.0.2
psycopg2-binary==2.9.6
python-decouple~=3.8
shopifyapi@ git+https://github.com/prikid/shopify_python_api@master
requests~=2.31.0
redis~=4.5.5
haggis~=0.9.1
python-dateutil~=2.8.2
celery-singleton
pyactiveresource~=2.2.2
django-celery-beat~=2.5.0
pydantic~=1.10.9

